pyproject.toml
[project]
name = "cora-customer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Niels", email = "niels.scholz01@web.de" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "gradio>=5.49.1",
    "ipykernel>=7.1.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-ollama>=1.0.0",
    "mcp[cli]>=1.21.0",
    "pyyaml>=6.0.3",
    "redisvl>=0.11.0",
    "sentence-transformers>=5.1.2",
]

[project.scripts]
run_mcp = "cora_customer_agent.mcp_server:main"
run_gradio = "cora_customer_agent.gradio_chatbot:main"
